PV2-1990 full test coverage for StoreDCJobStatus
